added BrowserInterface.php
